+++ packages/pkg-core/src/validation/__tests__/moon-query-projects.spec.ts
@@ -1,6 +1,7 @@
 import { addr } from "@business-as-code/address"
 import { xfs } from "@business-as-code/fslib"
 import { moonQueryProjects } from "../moon-query-projects"
+import { describe, it, jest, expect } from "bun:test";
 
 /**
  projectGraph here refers to the output files of moon at `.moon/cache/state/projectGraph.json`
@@ -57,15 +58,26 @@ describe('moon-query-projects', () => {
         expect(moonQueryProjects.parse(projectGraphJson)).toBeTruthy()
       })
     })
+    describe('1.10.1', () => {
+      it('validates a single-project moon projectGraph.json file', async () => {
+        const projectGraphJson = await xfs.readJsonPromise(addr.pathUtils.resolve(addr.parsePath(__dirname), addr.parsePath('./etc/query/projects/v1.10.1/language=typescript.json')).address)
+        expect(moonQueryProjects.parse(projectGraphJson)).toBeTruthy()
+      })
+      it('validates a full moon projectGraph.json file', async () => {
+        const projectGraphJson = await xfs.readJsonPromise(addr.pathUtils.resolve(addr.parsePath(__dirname), addr.parsePath('./etc/query/projects/v1.10.1/language=typescript-full.json')).address)
+        expect(moonQueryProjects.parse(projectGraphJson)).toBeTruthy()
+      })
+    })
   })
-  describe('1.10.1', () => {
+  describe('1.18.3', () => {
     it('validates a single-project moon projectGraph.json file', async () => {
-      const projectGraphJson = await xfs.readJsonPromise(addr.pathUtils.resolve(addr.parsePath(__dirname), addr.parsePath('./etc/query/projects/v1.10.1/language=typescript.json')).address)
+      const projectGraphJson = await xfs.readJsonPromise(addr.pathUtils.resolve(addr.parsePath(__dirname), addr.parsePath('./etc/query/projects/v1.18.3/language=typescript.json')).address)
       expect(moonQueryProjects.parse(projectGraphJson)).toBeTruthy()
     })
     it('validates a full moon projectGraph.json file', async () => {
-      const projectGraphJson = await xfs.readJsonPromise(addr.pathUtils.resolve(addr.parsePath(__dirname), addr.parsePath('./etc/query/projects/v1.10.1/language=typescript-full.json')).address)
+      const projectGraphJson = await xfs.readJsonPromise(addr.pathUtils.resolve(addr.parsePath(__dirname), addr.parsePath('./etc/query/projects/v1.18.3/language=typescript-full.json')).address)
       expect(moonQueryProjects.parse(projectGraphJson)).toBeTruthy()
     })
   })
+
 })