+++ packages/tests/pkg-tests-git-mock-server/CHANGELOG.md
@@ -1,5 +1,47 @@
 # git-http-mock-server
 
+## 0.0.0-latest-20231217101715
+
+### Patch Changes
+
+- Updated dependencies
+  - @business-as-code/core@0.0.0-latest-20231217101715
+
+## 0.0.0-latest-20231216222809
+
+### Patch Changes
+
+- Updated dependencies
+  - @business-as-code/core@0.0.0-latest-20231216222809
+
+## 0.0.0-latest-20231216222547
+
+### Patch Changes
+
+- Updated dependencies
+  - @business-as-code/core@0.0.0-latest-20231216222547
+
+## 0.0.0-latest-20231216220904
+
+### Patch Changes
+
+- Updated dependencies
+  - @business-as-code/core@0.0.0-latest-20231216220904
+
+## 0.0.0-latest-20231216220813
+
+### Patch Changes
+
+- Updated dependencies
+  - @business-as-code/core@0.0.0-latest-20231216220813
+
+## 0.0.0-latest-20231216220757
+
+### Patch Changes
+
+- Updated dependencies
+  - @business-as-code/core@0.0.0-latest-20231216220757
+
 ## 0.0.0-latest-20231119110156
 
 ### Patch Changes