+++ packages/tests/plugin-dev-tests/CHANGELOG.md
@@ -1,5 +1,71 @@
 # @business-as-code/plugin-dev-tests
 
+## 0.0.0-latest-20231217101715
+
+### Patch Changes
+
+- Making a release
+- Updated dependencies
+  - @business-as-code/address@0.0.0-latest-20231217101715
+  - @business-as-code/core@0.0.0-latest-20231217101715
+  - @business-as-code/error@0.0.0-latest-20231217101715
+  - @business-as-code/fslib@0.0.0-latest-20231217101715
+
+## 0.0.0-latest-20231216222809
+
+### Patch Changes
+
+- Making a release
+- Updated dependencies
+  - @business-as-code/address@0.0.0-latest-20231216222809
+  - @business-as-code/core@0.0.0-latest-20231216222809
+  - @business-as-code/error@0.0.0-latest-20231216222809
+  - @business-as-code/fslib@0.0.0-latest-20231216222809
+
+## 0.0.0-latest-20231216222547
+
+### Patch Changes
+
+- Making a release
+- Updated dependencies
+  - @business-as-code/address@0.0.0-latest-20231216222547
+  - @business-as-code/core@0.0.0-latest-20231216222547
+  - @business-as-code/error@0.0.0-latest-20231216222547
+  - @business-as-code/fslib@0.0.0-latest-20231216222547
+
+## 0.0.0-latest-20231216220904
+
+### Patch Changes
+
+- Making a release
+- Updated dependencies
+  - @business-as-code/address@0.0.0-latest-20231216220904
+  - @business-as-code/core@0.0.0-latest-20231216220904
+  - @business-as-code/error@0.0.0-latest-20231216220904
+  - @business-as-code/fslib@0.0.0-latest-20231216220904
+
+## 0.0.0-latest-20231216220813
+
+### Patch Changes
+
+- Making a release
+- Updated dependencies
+  - @business-as-code/address@0.0.0-latest-20231216220813
+  - @business-as-code/core@0.0.0-latest-20231216220813
+  - @business-as-code/error@0.0.0-latest-20231216220813
+  - @business-as-code/fslib@0.0.0-latest-20231216220813
+
+## 0.0.0-latest-20231216220757
+
+### Patch Changes
+
+- Making a release
+- Updated dependencies
+  - @business-as-code/address@0.0.0-latest-20231216220757
+  - @business-as-code/core@0.0.0-latest-20231216220757
+  - @business-as-code/error@0.0.0-latest-20231216220757
+  - @business-as-code/fslib@0.0.0-latest-20231216220757
+
 ## 0.0.0-latest-20231119110156
 
 ### Patch Changes